Removed direct database access from run.py
 - Removed `database_uri` parameter from RightmoveDatabase class.
 - Moved functions from run.py directly to database.py, cleaned up these functions.
